python/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.10.0+cpu
torchvision==0.25.0+cpu
onnx==1.20.1
onnxruntime==1.23.2
onnxscript==0.5.7
pocketbase==0.15.0
scikit-learn==1.8.0
skl2onnx==1.19.1
dagster==1.12.12
dagster-webserver==1.12.12
python-dotenv==1.2.1
